refactor(tests): split headless integration tests by feature
