+++ crates/arf-console/tests/headless/mod.rs
@@ -1,0 +1,7 @@
+mod error_codes;
+mod eval;
+mod history;
+mod lifecycle_flags;
+mod output_encoding;
+mod platform;
+mod support;
